fix: Change the way jwt is created and handle it by the services
