+++ app/internal/core/middlewares/configurations.js
@@ -1,0 +1,6 @@
+import { defaultConfig } from '../providers';
+
+export const configurations = (req, res, next) => {
+  req.config = defaultConfig;
+  next();
+};
