+++ app/internal/core/providers/config.provider.js
@@ -1,0 +1,5 @@
+import 'dotenv/config';
+
+export const defaultConfig = {
+  security_key: process.env.SECRET_KEY,
+};
