+++ app/internal/core/providers/index.js
@@ -1,0 +1,1 @@
+export * from './config.provider';
