+++ app/server/index.js
@@ -2,7 +2,8 @@ import bodyParser from 'body-parser';
 import cors from 'cors';
 import express from 'express';
 import { Adapters } from '../internal/adapters';
-import { Logger, HandlerError, VerifyToken } from '../internal/core/middlewares';
+import { VerifyToken } from '../internal/adapters/auth';
+import { Logger, HandlerError, configurations } from '../internal/core/middlewares';
 
 const { Routes } = Adapters;
 
@@ -11,6 +12,7 @@ App.use(cors());
 App.use(bodyParser.json());
 App.use(bodyParser.urlencoded({ extended: false }));
 App.use(Logger);
+App.use(configurations);
 
 Routes.MainRoute(App);
 Routes.AuthRoute(App);
